Get list of links to IATF resolutions from Philippines Officie Gazette

Get list of links to IATF resolutions from Philippines Officie Gazette

get_iatf_gazette(base, pages = 1:5)

Arguments

base
URL to the IATF resolutions webpage in the Official Gazette. This is currently at https://www.officialgazette.gov.ph/section/laws/other-issuances/inter-agency-task-force-for-the-management-of-emerging-infectious-diseases-resolutions/

pages
A vector of paginated webpages in which the IATF resolutions are available. This currently defaults to 1:5 as there are currenlty 5 pages.

Value

A tibble containing absolute links to all the current IATF resolutions at time of extraction from the Official Gazette. The tibble contains the following information:

idResolution number

titleTitle of resolution

dateDate (in <YYYY-MM-DD> format) resolution was issued

sourceSource of resolution. This is by default from IATF

typeType of document. This is by default a resolution

urlAbsolute URL for PDF of resolution

checkedDate (in <YYYY-MM-DD format) table was extracted. This is by default provided by Sys.Date()

Examples

base <- "https://www.officialgazette.gov.ph/section/laws/other-issuances"
agency <- "inter-agency-task-force-for-the-management-of-emerging-infectious-diseases-resolutions/"

get_iatf_gazette(base = paste(base, agency, sep = "/"), pages = 1)

#> # A tibble: 10 x 7
#>       id title              date       source type  url               checked   
#>    <dbl> <chr>              <date>     <chr>  <chr> <chr>             <date>    
<<<<<<< HEAD
#>  1    46 RESOLUTION NO. 46… 2020-06-15 IATF   reso… https://www.offi… 2020-06-18
#>  2    46 RESOLUTION NO. 46… 2020-06-15 IATF   reso… https://www.offi… 2020-06-18
#>  3    45 RESOLUTION NO. 45… 2020-06-10 IATF   reso… https://www.offi… 2020-06-18
#>  4    44 RESOLUTION NO. 44… 2020-06-08 IATF   reso… https://www.offi… 2020-06-18
#>  5    NA OMNIBUS GUIDELINE… 2020-06-03 IATF   reso… https://www.offi… 2020-06-18
#>  6    43 RESOLUTION NO. 43… 2020-06-03 IATF   reso… https://www.offi… 2020-06-18
#>  7    42 RESOLUTION NO. 42… 2020-06-01 IATF   reso… https://www.offi… 2020-06-18
#>  8    41 RESOLUTION NO. 41… 2020-05-29 IATF   reso… https://www.offi… 2020-06-18
#>  9    40 RESOLUTION NO. 40… 2020-05-27 IATF   reso… https://www.offi… 2020-06-18
#> 10    NA OMNIBUS GUIDELINE… 2020-05-22 IATF   reso… https://www.offi… 2020-06-18

=======
#>  1    50 RESOLUTION NO. 50… 2020-06-29 IATF   reso… https://www.offi… 2020-06-30
#>  2    NA OMNIBUS GUIDELINE… 2020-06-25 IATF   reso… https://www.offi… 2020-06-30
#>  3    49 RESOLUTION NO. 49… 2020-06-25 IATF   reso… https://www.offi… 2020-06-30
#>  4    48 RESOLUTION NO. 48… 2020-06-22 IATF   reso… https://www.offi… 2020-06-30
#>  5    47 RESOLUTION NO. 47… 2020-06-19 IATF   reso… https://www.offi… 2020-06-30
#>  6    46 RESOLUTION NO. 46… 2020-06-15 IATF   reso… https://www.offi… 2020-06-30
#>  7    46 RESOLUTION NO. 46… 2020-06-15 IATF   reso… https://www.offi… 2020-06-30
#>  8    45 RESOLUTION NO. 45… 2020-06-10 IATF   reso… https://www.offi… 2020-06-30
#>  9    44 RESOLUTION NO. 44… 2020-06-08 IATF   reso… https://www.offi… 2020-06-30
#> 10    NA OMNIBUS GUIDELINE… 2020-06-03 IATF   reso… https://www.offi… 2020-06-30

>>>>>>> 70172fc196460ac72e2c31adf15dc960e841e6b9